changed the element type to be different from the index type to bring out more errors
git-svn-id: http://svn.boost.org/svn/boost/trunk@21771 b8fc166d-592f-0410-95f2-cb63ce0dd405

+++ boost/libs/iterator/test/permutation_iterator_test.cpp
@@ -20,7 +20,7 @@
 void permutation_test()
 {
   // Example taken from documentation of old permutation_iterator.
-  typedef std::vector< int > element_range_type;
+  typedef std::vector< double > element_range_type;
   typedef std::list< int > index_type;
 
   const int element_range_size = 10;
